Add news_article+page_view_model.dart
