+++ lib/views/pages/newsfeed/news_article.dart
@@ -16,6 +16,8 @@ import 'package:talawa/utils/custom_toast.dart';
 import 'package:talawa/utils/gql_client.dart';
 import 'package:talawa/utils/api_functions.dart';
 import 'package:talawa/utils/uidata.dart';
+import 'package:talawa/view_models/newwfeed_view_model/new_article_page_view_model.dart';
+import 'package:talawa/views/base_view.dart';
 
 const String newLineKey = "@123TALAWA321@";
 
@@ -38,85 +40,23 @@ class _NewsArticleState extends State<NewsArticle> {
     }
   }
 
-  TextEditingController commentController;
-  Preferences preferences = Preferences();
-  ApiFunctions apiFunctions = ApiFunctions();
-  bool showLoadComments = false;
-  List comments = [];
-  bool moreComments = false;
-  bool isCommentAdded = false;
-  int index;
-  Map post;
-  final Queries _query = Queries();
-  List userDetails = [];
-  String userID;
-  String orgName;
-  GraphQLConfiguration graphQLConfiguration = GraphQLConfiguration();
-  PostController postController;
-
-  @override
-  void initState() {
-    super.initState();
-    commentController = TextEditingController(
-        text: Provider.of<CommentHandler>(context, listen: false)
-            .comment(widget.post["_id"].toString()));
-    fetchUserDetails();
-    index = widget.index;
-    post = widget.post;
-    commentController.addListener(_notifyData);
-  }
-
-  void _notifyData() {
-    Provider.of<CommentHandler>(context, listen: false)
-        .commentEntry(widget.post["_id"].toString(), commentController.text);
-  }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    getPostComments();
-  }
-
-  Future fetchUserDetails() async {
-    userID = await preferences.getUserId();
-    final GraphQLClient _client = graphQLConfiguration.clientToQuery();
-    final QueryResult result = await _client.query(QueryOptions(
-        documentNode: gql(_query.fetchUserInfo), variables: {'id': userID}));
-    if (result.hasException) {
-      print(result.exception);
-      CustomToast.exceptionToast(msg: result.exception.toString());
-    } else if (!result.hasException) {
-      //print(result);
-      setState(() {
-        userDetails = result.data['users'] as List;
-      });
-      //print(userDetails);
-    }
-  }
-
-  @override
-  void dispose() {
-    commentController.dispose();
-    super.dispose();
-  }
-
-//return profile image of current user
-  Widget _profileImage() {
-    return userDetails[0]['image'] != null
+  //return profile image of current user
+  Widget _profileImage(NewsArticleViewModel model) {
+    return model.userDetails[0]['image'] != null
         ? CircleAvatar(
             radius: 30,
             backgroundImage: NetworkImage(
                 Provider.of<GraphQLConfiguration>(context).displayImgRoute +
-                    userDetails[0]['image'].toString()))
+                    model.userDetails[0]['image'].toString()))
         : CircleAvatar(
             radius: 45.0,
             backgroundColor: Colors.white,
             child: Text(
-                userDetails[0]['firstName']
+                model.userDetails[0]['firstName']
                         .toString()
                         .substring(0, 1)
                         .toUpperCase() +
-                    userDetails[0]['lastName']
+                    model.userDetails[0]['lastName']
                         .toString()
                         .substring(0, 1)
                         .toUpperCase(),
@@ -126,111 +66,15 @@ class _NewsArticleState extends State<NewsArticle> {
           );
   }
 
-  //this method helps us to get the comments of the post
-  Future getPostComments() async {
-    final String mutation =
-        Queries().getPostsComments(widget.post['_id'].toString());
-    final Map result = await apiFunctions.gqlmutation(mutation) as Map;
-    comments = result == null
-        ? []
-        : (result['commentsByPost'] as List).reversed.toList();
-  }
-
-  //this method helps us to create any comments we are willing to
-  Future createComment() async {
-    FocusScope.of(context).unfocus();
-    String queryText = '';
-    if (commentController.text.isNotEmpty) {
-      Fluttertoast.showToast(msg: "Adding Comment...");
-      queryText = commentController.text.replaceAll("\n", newLineKey).trim();
-      final Map result = await Queries()
-          .createComments(widget.post['_id'].toString(), queryText) as Map;
-      if (result == null) {
-        Fluttertoast.showToast(
-          msg: "Sorry, this comment could not be posted.",
-        );
-      } else {
-        isCommentAdded = true;
-        FocusScope.of(context).requestFocus(FocusNode());
-        commentController.text = '';
-        await Fluttertoast.showToast(
-          msg: "Comment added.",
-        );
-        postController.addComment(index, result["createComment"] as Map);
-      }
-    } else {
-      Fluttertoast.showToast(msg: "Please write comment");
-    }
-  }
-
-  //get time of comment
-  String commentTime(int index) {
-    final Duration commentTimeDuration = DateTime.now().difference(
-      DateTime.fromMillisecondsSinceEpoch(
-          int.parse(comments[index]['createdAt'].toString())),
-    );
-
-    String timeText = '';
-
-    if (commentTimeDuration.inMinutes < 1) {
-      if (commentTimeDuration.inSeconds == 1) {
-        timeText = ' second ago';
-      } else {
-        timeText = ' seconds ago';
-      }
-      return commentTimeDuration.inSeconds.toString() + timeText;
-    } else if (commentTimeDuration.inHours < 1) {
-      if (commentTimeDuration.inMinutes == 1) {
-        timeText = ' min ago';
-      } else {
-        timeText = ' mins ago';
-      }
-      return commentTimeDuration.inMinutes.toString() + timeText;
-    } else if (commentTimeDuration.inDays < 1) {
-      if (commentTimeDuration.inHours == 1) {
-        timeText = ' hour ago';
-      } else {
-        timeText = ' hours ago';
-      }
-      return commentTimeDuration.inHours.toString() + timeText;
-    } else if (commentTimeDuration.inDays < 7) {
-      if (commentTimeDuration.inDays == 1) {
-        timeText = ' day ago';
-      } else {
-        timeText = ' days ago';
-      }
-      return commentTimeDuration.inDays.toString() + timeText;
-    } else if (commentTimeDuration.inDays < 52) {
-      final int weeks = commentTimeDuration.inDays ~/ 7;
-      if (weeks == 1) {
-        timeText = ' week ago';
-      } else {
-        timeText = ' weeks ago';
-      }
-      return weeks.toString() + timeText;
-    } else {
-      final int years = commentTimeDuration.inDays ~/ 365;
-      if (years == 1) {
-        timeText = ' year ago';
-      } else {
-        timeText = ' years ago';
-      }
-      return years.toString() + timeText;
-    }
-  }
-
-  String addNewline(String rawComment) {
-    // ignore: parameter_assignments
-    rawComment = rawComment.replaceAll(newLineKey, "\n");
-    return rawComment;
-  }
-
   //main build starts here
   @override
   Widget build(BuildContext context) {
-    return WillPopScope(
+    return BaseView<NewsArticleViewModel>(
+      onModelReady: (model) =>
+          model.initialize(widget.post, widget.index, context),
+      builder: (context, model, child) => WillPopScope(
         onWillPop: () async {
-          Navigator.of(context).pop(isCommentAdded);
+          Navigator.of(context).pop(model.isCommentAdded);
           return true;
         },
         child: Scaffold(
@@ -240,7 +84,7 @@ class _NewsArticleState extends State<NewsArticle> {
             elevation: 0.0,
             leading: GestureDetector(
               onTap: () {
-                Navigator.of(context).pop(isCommentAdded);
+                Navigator.of(context).pop(model.isCommentAdded);
               },
               child: const Icon(
                 Icons.arrow_back,
@@ -300,7 +144,9 @@ class _NewsArticleState extends State<NewsArticle> {
                     Flexible(
                       flex: 3,
                       child: ListTile(
-                        leading: userDetails.isEmpty ? null : _profileImage(),
+                        leading: model.userDetails.isEmpty
+                            ? null
+                            : _profileImage(model),
                         title: Container(
                           constraints: const BoxConstraints(
                             maxHeight: double.infinity,
@@ -331,8 +177,8 @@ class _NewsArticleState extends State<NewsArticle> {
                                 color: Colors.grey,
                                 icon: const Icon(Icons.send),
                                 onPressed: () {
-                                  print(commentController.text);
-                                  createComment();
+                                  print(model.commentController.text);
+                                  model.createComment();
                                 },
                               ),
                               hintText: 'Leave a Comment...',
@@ -343,7 +189,7 @@ class _NewsArticleState extends State<NewsArticle> {
                                 ),
                               ),
                             ),
-                            controller: commentController,
+                            controller: model.commentController,
                           ),
                         ),
                       ),
@@ -351,49 +197,44 @@ class _NewsArticleState extends State<NewsArticle> {
                     Flexible(
                       flex: 10,
                       child: Container(
-                          child: showLoadComments == false
+                          child: model.showLoadComments == false
                               ? Align(
                                   alignment: Alignment.topCenter,
-                                  child: loadCommentsButton())
-                              : commentList()),
+                                  child: loadCommentsButton(model))
+                              : commentList(model)),
                     ),
                   ],
                 ),
               ),
             ],
           ),
-        ));
+        ),
+      ),
+    );
   }
 
   //this loads the comments button
-  Widget loadCommentsButton() {
+  Widget loadCommentsButton(NewsArticleViewModel model) {
     return TextButton(
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all<Color>(Colors.grey[200]),
-        ),
-        onPressed: () {
-          setState(() {
-            showLoadComments = true;
-          });
-        },
-        child: const Text(
-          'Load Comments',
-          style: TextStyle(color: Colors.black54),
-        ));
-  }
-
-  // For getting length of Comments to be displayed
-  int getCommentslength() {
-    getPostComments();
-    return comments.length;
+      style: ButtonStyle(
+        backgroundColor: MaterialStateProperty.all<Color>(Colors.grey[200]),
+      ),
+      onPressed: () {
+        model.changeLoading(value: true);
+      },
+      child: const Text(
+        'Load Comments',
+        style: TextStyle(color: Colors.black54),
+      ),
+    );
   }
 
   // a new widget for comment list
-  Widget commentList() {
-    int lenthOfCommentList = getCommentslength();
+  Widget commentList(NewsArticleViewModel model) {
+    int lenthOfCommentList = model.getCommentslength();
 
     if (lenthOfCommentList > 3) {
-      if (moreComments == false) {
+      if (model.moreComments == false) {
         lenthOfCommentList = 3;
       }
     }
@@ -403,7 +244,7 @@ class _NewsArticleState extends State<NewsArticle> {
         ListTile(
           key: const ValueKey('commentIcon'),
           leading: const Icon(Icons.chat),
-          title: Text('${comments.length}  Comments'),
+          title: Text('${model.comments.length}  Comments'),
         ),
         Flexible(
           child: ListView.builder(
@@ -411,6 +252,8 @@ class _NewsArticleState extends State<NewsArticle> {
               physics: const ClampingScrollPhysics(),
               itemCount: lenthOfCommentList,
               itemBuilder: (context, index) {
+                final Map<String, String> creator =
+                    model.comments[index]['creator'] as Map<String, String>;
                 return ListTile(
                   leading: const CircleAvatar(
                     backgroundColor: UIData.secondaryColor,
@@ -420,34 +263,31 @@ class _NewsArticleState extends State<NewsArticle> {
                     ),
                   ),
                   title: Text(
-                    comments[index]['text'].toString(),
+                    model.comments[index]['text'].toString(),
                   ),
                   subtitle: Row(
                     children: [
-                      Text(
-                          '${comments[index]['creator']['firstName']} ${comments[index]['creator']['lastName']}'),
+                      Text('${creator['firstName']} ${creator['lastName']}'),
                       const Text(
                         " - ",
                         style: TextStyle(
                           fontSize: 20,
                         ),
                       ),
-                      Text(commentTime(index)),
+                      Text(model.commentTime(index)),
                     ],
                   ),
                 );
               }),
         ),
-        (moreComments || comments.length <= 3)
+        (model.moreComments || model.comments.length <= 3)
             ? const SizedBox(
                 width: 0,
                 height: 0,
               )
             : TextButton(
                 onPressed: () {
-                  setState(() {
-                    moreComments = true;
-                  });
+                  model.showMoreComments(value: true);
                 },
                 child: const Text("View More Comments"))
       ],
